Refactor : MVC 패턴에 맞게 패키지 구조 변경
