+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import controller.LottoController;
+import lotto.controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
