+++ src/main/java/lotto/controller/LottoController.java
@@ -1,19 +1,19 @@
-package controller;
+package lotto.controller;
 
-import constants.Grade;
+import lotto.constants.Grade;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.WinningInformation;
-import lotto.Lotto;
-import lotto.LottoNumber;
-import lotto.Money;
-import lotto.PointCalculator;
-import lotto.PointResult;
-import utils.Parser;
-import utils.RandomUtils;
-import view.InputView;
-import view.OutputView;
+import lotto.domain.WinningInformation;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+import lotto.domain.Money;
+import lotto.domain.PointCalculator;
+import lotto.domain.PointResult;
+import lotto.utils.Parser;
+import lotto.utils.RandomUtils;
+import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class LottoController {
     public void run() {
