+++ src/main/java/lotto/domain/Lotto.java
@@ -1,7 +1,7 @@
-package lotto;
+package lotto.domain;
 
-import constants.ErrorMessage;
-import constants.NumberType;
+import lotto.constants.ErrorMessage;
+import lotto.constants.NumberType;
 import java.util.List;
 
 public class Lotto {
