+++ src/main/java/lotto/domain/Money.java
@@ -1,7 +1,7 @@
-package lotto;
+package lotto.domain;
 
-import constants.ErrorMessage;
-import constants.NumberType;
+import lotto.constants.ErrorMessage;
+import lotto.constants.NumberType;
 
 public class Money {
