+++ src/main/java/lotto/utils/Parser.java
@@ -1,8 +1,8 @@
-package utils;
+package lotto.utils;
 
 import java.util.Arrays;
 import java.util.List;
-import lotto.LottoNumber;
+import lotto.domain.LottoNumber;
 
 public class Parser {
     public static List<LottoNumber> stringToLottoNumbers(String userInput) {
