+++ src/test/java/ApplicationTest.java
@@ -1,6 +1,5 @@
-package lotto;
-
 import camp.nextstep.edu.missionutils.test.NsTest;
+import lotto.Application;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
