+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
-import constants.ErrorMessage;
-import constants.NumberType;
+import lotto.constants.ErrorMessage;
+import lotto.constants.NumberType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
